Switch to 'fancy' versioning for Enterprise RPMs
Similar to community editions, we are switching to 'fancy' versioning
for Citus Enterprise.

Merging this will release citus-enterprise60_95/citus-enterprise60_96.
The version will be 6.0.1.citus-2 and existing users will be auto
upgraded. This and all subsequent 'fancy' version packages will
conflict with one another, ensuring two versions cannot be installed
at the same time.

The one slight change from Community is that these packages provide and
conflict with e.g. citus_96 in order to correctly conflict with the OSS
packages.