Add Debian Stretch to OS list; fix gpg problem
So this is pretty complex, but basically Debian Stretch no longer comes
with gpg installed, which is now just a "Recommend" dependency of apt.
The distinction between gpg and gpg2 is often obscured depending upon
the OS flavor, and furthermore, gpg2 unbundles dirmngr, which is good
to have for our build process.

The only reason all of this is necessary at all is our invocation of
apt-key. For a little bit I mistakenly believed that using apt-key add
with a local file rather than using a keyserver might avoid the need
for dirmngr in particular, but after finding some discussion online I
decided to go with this approach.